requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
PyMySQL==1.1.1
requests==2.32.3
Werkzeug==3.1.3
APScheduler==3.10.4
captcha==0.7.1
pycryptodome>=3.19.1
python-dotenv>=1.0.0
gunicorn>=21.2.0
